further improve user menu
