+++ packages/client-core/src/user/components/UserMenu/menus/SettingMenu.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React, { useState } from 'react'
 import { Mic, VolumeUp, BlurLinear } from '@mui/icons-material'
 import Checkbox from '@mui/material/Checkbox'
 import FormControlLabel from '@mui/material/FormControlLabel'
@@ -9,16 +9,22 @@ import { EngineRendererAction, useEngineRendererState } from '@xrengine/engine/s
 import { useTranslation } from 'react-i18next'
 import { UserSetting } from '@xrengine/common/src/interfaces/User'
 import { dispatchLocal } from '@xrengine/engine/src/networking/functions/dispatchFrom'
-import { EngineRenderer } from '@xrengine/engine/src/renderer/WebGLRendererSystem'
+import { AuthService, useAuthState } from '../../../services/AuthService'
 
-type Props = {
-  setting: UserSetting
-  setUserSettings: (settings: Partial<UserSetting>) => void
-}
-
-const SettingMenu = (props: Props): JSX.Element => {
+const SettingMenu = (): JSX.Element => {
   const { t } = useTranslation()
   const rendererState = useEngineRendererState()
+
+  const authState = useAuthState()
+  const selfUser = authState.user
+  const [userSettings, setUserSetting] = useState<UserSetting>(selfUser?.user_setting.value!)
+
+  const setUserSettings = (newSetting: any): void => {
+    const setting = { ...userSettings, ...newSetting }
+    setUserSetting(setting)
+    AuthService.updateUserSettings(selfUser.user_setting.value?.id, setting)
+  }
+
   return (
     <div className={styles.menuPanel}>
       <div className={styles.settingPanel}>
@@ -32,9 +38,9 @@ const SettingMenu = (props: Props): JSX.Element => {
             </span>
             <span className={styles.settingLabel}>{t('user:usermenu.setting.lbl-volume')}</span>
             <Slider
-              value={props.setting?.volume == null ? 100 : props.setting?.volume}
+              value={userSettings?.volume == null ? 100 : userSettings?.volume}
               onChange={(_, value: number) => {
-                props.setUserSettings({ volume: value })
+                setUserSettings({ volume: value })
                 const mediaElements = document.querySelectorAll<HTMLMediaElement>('video, audio')
                 for (let i = 0; i < mediaElements.length; i++) {
                   mediaElements[i].volume = (value as number) / 100
@@ -51,9 +57,9 @@ const SettingMenu = (props: Props): JSX.Element => {
             </span>
             <span className={styles.settingLabel}>{t('user:usermenu.setting.lbl-microphone')}</span>
             <Slider
-              value={props.setting?.microphone == null ? 100 : props.setting?.microphone}
+              value={userSettings?.microphone == null ? 100 : userSettings?.microphone}
               onChange={(_, value: number) => {
-                props.setUserSettings({ microphone: value })
+                setUserSettings({ microphone: value })
               }}
               className={styles.slider}
               max={100}
